game: move damage routines into the moves.ts file
